requirements.txt
Flask
Flask-CORS
openai-whisper
sounddevice
numpy
ffmpeg-python
websockets